ompl_setup: move GridMap to own module
